refactor : NotCountScore 예외처리 추가 및 테스트 코드 변경
