+++ src/main/java/model/frame/score/NotCountScore.java
@@ -1,0 +1,7 @@
+package model.frame.score;
+
+public class NotCountScore extends RuntimeException {
+    public NotCountScore(String errorMessage) {
+        super(errorMessage);
+    }
+}
